requirement.txt
django
Pillow
psycopg2-binary
djangorestframework
